files

+++ coq/clients/tags/sql/create/tables.sql
@@ -2,9 +2,9 @@ BEGIN;
 
 
 CREATE TABLE IF NOT EXISTS files (
-  filename TEXT    NOT NULL PRIMARY KEY,
-  filetype TEXT    NOT NULL,
-  mtime    INTEGER NOT NULL,
+  filename TEXT NOT NULL PRIMARY KEY,
+  filetype TEXT NOT NULL,
+  mtime    REAL NOT NULL,
 ) WITHOUT ROWID;
 CREATE INDEX IF NOT EXISTS files_filetype ON files (filetype);
